El juego se compilaba SIN OPTIMIZAR, y dos generadores de textura tenian comportamiento indefinido.
CMakeLists.txt no fijaba CMAKE_BUILD_TYPE ni ninguna bandera -O, asi que psp-g++
compilaba en -O0: el juego entero corria sin optimizar en una CPU MIPS de 333 MHz,
con colision, gravedad, poses y audio procedural hechos a base de coma flotante.
Medido: el EBOOT baja de 1.460.000 a 257.160 bytes, un 82% menos, porque -O2
ademas elimina el codigo muerto de los mundos viejos (catedrales, torres, puentes)
que -O0 arrastraba entero dentro del binario.

-O2 y no -O3 a proposito: -O3 daba 329 KB, un 28% mas de codigo, y la PSP tiene
solo 16 KB de cache de instrucciones. Ahi el desenrollado agresivo cuesta mas
fallos de cache de lo que ahorra en instrucciones.

Y antes de encender la optimizacion habia que desactivar una mina. El hash entero
de las texturas hacia `x * 374761393` con x de tipo int: a partir de x=6 eso
DESBORDA el entero con signo, que en C++ es comportamiento indefinido, no un
envoltorio garantizado. A -O0 se envolvia por casualidad y salia bien; con -O2 el
compilador tiene permiso para asumir que nunca ocurre y generar cualquier cosa.
GCC lo estaba avisando en dos sitios: la textura del SUELO (ground_tex.h) y la de
las ventanas (agent_textures.h).

Arreglado en los cuatro sitios que lo tenian (ground_tex.h, agent_textures.h,
gothic_tex.h, gothic_bldg.h) casteando a unsigned ANTES de multiplicar.
facade_tex.h e industrial_tex.h ya estaban bien escritos y no se tocaron.
La textura sale bit a bit identica: en complemento a dos la multiplicacion con
signo y la sin signo dan los mismos bits, asi que esto no cambia un solo pixel,
solo deja de ser codigo que el compilador puede romper cuando le convenga.

+++ src/gothic_bldg.h
@@ -32,8 +32,8 @@ static void buildGothicBldg(TexVertex *buf, int &i, float cx, float cz,
     const bool  longX = (w >= d);                                         // eje LARGO (nave) = X? (E/O -> false)
 
     // ---- variacion determinista por edificio (hash de cx,cz,h; NADA de rand) ----
-    unsigned int hsh = (unsigned int)((int)cx * 374761393) + (unsigned int)((int)cz * 668265263)
-                     + (unsigned int)((int)h * (int)2654435761u);
+    unsigned int hsh = (unsigned int)(int)cx * 374761393u + (unsigned int)(int)cz * 668265263u
+                     + (unsigned int)(int)h * 2654435761u;
     hsh ^= hsh >> 15; hsh *= 2246822519u; hsh ^= hsh >> 13; hsh *= 3266489917u; hsh ^= hsh >> 16;
     const float v0 = (float)(hsh & 0xFFu)        * (1.0f / 255.0f);
     const float v1 = (float)((hsh >> 8)  & 0xFFu) * (1.0f / 255.0f);
